Update README.md
Updated the chapters 1-2-3. Sorry that I forgot to comment on my previous commit. 

@aliasimon, fyi I first destroyed then recovered the changes you recently added. Sorry about that! Now it should be fine.

In general, I believe we should reduce the # of content in chapter 1, so that we will mention each document only one time. 

I have two suggestions: 
a) We can discuss which document belongs to which chapter and separate them.
b) We can discuss whether chapter 1&2 can be merged under a common name such as "Gather Knowledge & Design Architecture".